+++ packages/rocketh-deploy/src/index.ts
@@ -1,5 +1,5 @@
 import {Abi} from 'abitype';
-import {EIP1193DATA, EIP1193TransactionData} from 'eip-1193';
+import {EIP1193TransactionData} from 'eip-1193';
 import type {
 	Artifact,
 	DeploymentConstruction,
@@ -7,36 +7,18 @@ import type {
 	Environment,
 	PendingDeployment,
 	PartialDeployment,
-	PendingExecution,
 	Signer,
 } from 'rocketh';
 import {extendEnvironment} from 'rocketh';
-import {
-	Address,
-	Chain,
-	ContractFunctionArgs,
-	ContractFunctionName,
-	DecodeFunctionResultReturnType,
-	EncodeDeployDataParameters,
-	ReadContractParameters,
-	WriteContractParameters,
-	decodeFunctionResult,
-	encodeFunctionData,
-	encodePacked,
-	keccak256,
-} from 'viem';
-import {DeployContractParameters, encodeDeployData} from 'viem';
+import {Address, Chain, encodePacked, keccak256} from 'viem';
+import {encodeDeployData} from 'viem';
 import {logs} from 'named-logs';
 
-const logger = logs('rocketh-deploy');
+const logger = logs('@rocketh/deploy');
 
 declare module 'rocketh' {
 	interface Environment {
 		deploy: DeployFunction;
-		execute: ExecuteFunction;
-		read: ReadFunction;
-		executeByName: ExecuteFunctionByName;
-		readByName: ReadFunctionByName;
 	}
 }
 
@@ -45,82 +27,6 @@ export type DeployFunction = <TAbi extends Abi, TChain extends Chain = Chain>(
 	args: DeploymentConstruction<TAbi>,
 	options?: DeployOptions
 ) => Promise<Deployment<TAbi> & {updated: boolean}>;
-
-export type ExecuteFunction = <
-	TAbi extends Abi,
-	TFunctionName extends ContractFunctionName<TAbi, 'nonpayable' | 'payable'>,
-	TArgs extends ContractFunctionArgs<TAbi, 'nonpayable' | 'payable', TFunctionName> = ContractFunctionArgs<
-		TAbi,
-		'nonpayable' | 'payable',
-		TFunctionName
-	>
->(
-	deployment: Deployment<TAbi>,
-	args: ExecutionArgs<TAbi, TFunctionName, TArgs>
-) => Promise<EIP1193DATA>;
-
-export type ExecuteFunctionByName = <
-	TAbi extends Abi,
-	TFunctionName extends ContractFunctionName<TAbi, 'nonpayable' | 'payable'>,
-	TArgs extends ContractFunctionArgs<TAbi, 'nonpayable' | 'payable', TFunctionName> = ContractFunctionArgs<
-		TAbi,
-		'nonpayable' | 'payable',
-		TFunctionName
-	>
->(
-	name: string,
-	args: ExecutionArgs<TAbi, TFunctionName, TArgs>
-) => Promise<EIP1193DATA>;
-
-export type ReadFunction = <
-	TAbi extends Abi,
-	TFunctionName extends ContractFunctionName<TAbi, 'pure' | 'view'>,
-	TArgs extends ContractFunctionArgs<TAbi, 'pure' | 'view', TFunctionName> = ContractFunctionArgs<
-		TAbi,
-		'pure' | 'view',
-		TFunctionName
-	>
->(
-	deployment: Deployment<TAbi>,
-	args: ReadingArgs<TAbi, TFunctionName, TArgs>
-) => Promise<DecodeFunctionResultReturnType<TAbi, TFunctionName>>;
-
-export type ReadFunctionByName = <
-	TAbi extends Abi,
-	TFunctionName extends ContractFunctionName<TAbi, 'pure' | 'view'>,
-	TArgs extends ContractFunctionArgs<TAbi, 'pure' | 'view', TFunctionName> = ContractFunctionArgs<
-		TAbi,
-		'pure' | 'view',
-		TFunctionName
-	>
->(
-	name: string,
-	args: ReadingArgs<TAbi, TFunctionName, TArgs>
-) => Promise<DecodeFunctionResultReturnType<TAbi, TFunctionName>>;
-
-export type ExecutionArgs<
-	TAbi extends Abi,
-	TFunctionName extends ContractFunctionName<TAbi, 'nonpayable' | 'payable'>,
-	TArgs extends ContractFunctionArgs<TAbi, 'nonpayable' | 'payable', TFunctionName> = ContractFunctionArgs<
-		TAbi,
-		'nonpayable' | 'payable',
-		TFunctionName
-	>
-> = Omit<WriteContractParameters<TAbi, TFunctionName, TArgs>, 'address' | 'abi' | 'account' | 'nonce' | 'chain'> & {
-	account: string;
-};
-
-export type ReadingArgs<
-	TAbi extends Abi,
-	TFunctionName extends ContractFunctionName<TAbi, 'pure' | 'view'>,
-	TArgs extends ContractFunctionArgs<TAbi, 'pure' | 'view', TFunctionName> = ContractFunctionArgs<
-		TAbi,
-		'pure' | 'view',
-		TFunctionName
-	>
-> = Omit<ReadContractParameters<TAbi, TFunctionName, TArgs>, 'address' | 'abi' | 'account' | 'nonce'> & {
-	account?: string;
-};
 
 export type DeployOptions = {
 	linkedData?: any;
@@ -222,182 +128,6 @@ function linkLibraries(
 }
 
 extendEnvironment((env: Environment) => {
-	async function execute<
-		TAbi extends Abi,
-		TFunctionName extends ContractFunctionName<TAbi, 'nonpayable' | 'payable'>,
-		TArgs extends ContractFunctionArgs<TAbi, 'nonpayable' | 'payable', TFunctionName> = ContractFunctionArgs<
-			TAbi,
-			'nonpayable' | 'payable',
-			TFunctionName
-		>
-	>(deployment: Deployment<TAbi>, args: ExecutionArgs<TAbi, TFunctionName, TArgs>) {
-		const {account, ...viemArgs} = args;
-		let address: `0x${string}`;
-		if (account.startsWith('0x')) {
-			address = account as `0x${string}`;
-		} else {
-			if (env.namedAccounts) {
-				address = env.namedAccounts[account];
-				if (!address) {
-					throw new Error(`no address for ${account}`);
-				}
-			} else {
-				throw new Error(`no accounts setup, cannot get address for ${account}`);
-			}
-		}
-
-		const artifactToUse = deployment as unknown as Artifact<TAbi>;
-		const abi = artifactToUse.abi;
-		const calldata = encodeFunctionData<TAbi, TFunctionName>({
-			abi,
-			functionName: viemArgs.functionName,
-			args: viemArgs.args,
-		} as any);
-
-		const signer = env.addressSigners[address];
-
-		const txParam: EIP1193TransactionData = {
-			to: deployment.address,
-			type: '0x2',
-			from: address,
-			chainId: `0x${env.network.chain.id.toString(16)}` as `0x${string}`,
-			data: calldata,
-			gas: viemArgs.gas && (`0x${viemArgs.gas.toString(16)}` as `0x${string}`),
-			// gasPrice: viemArgs.gasPrice && `0x${viemArgs.gasPrice.toString(16)}` as `0x${string}`,
-			maxFeePerGas: viemArgs.maxFeePerGas && (`0x${viemArgs.maxFeePerGas.toString(16)}` as `0x${string}`),
-			maxPriorityFeePerGas:
-				viemArgs.maxPriorityFeePerGas && (`0x${viemArgs.maxPriorityFeePerGas.toString(16)}` as `0x${string}`),
-			// nonce: viemArgs.nonce && (`0x${viemArgs.nonce.toString(16)}` as `0x${string}`),
-		};
-		if (viemArgs.value) {
-			txParam.value = `0x${viemArgs.value?.toString(16)}` as `0x${string}`;
-		}
-
-		let txHash: `0x${string}`;
-		if (signer.type === 'wallet' || signer.type === 'remote') {
-			txHash = await signer.signer.request({
-				method: 'eth_sendTransaction',
-				params: [txParam] as any, // TODO fix eip-1193 ?,,
-			});
-		} else {
-			const rawTx = await signer.signer.request({
-				method: 'eth_signTransaction',
-				params: [txParam],
-			});
-
-			txHash = await env.network.provider.request({
-				method: 'eth_sendRawTransaction',
-				params: [rawTx],
-			});
-		}
-
-		const pendingExecution: PendingExecution = {
-			type: 'execution',
-			transaction: {hash: txHash, origin: address},
-			// description, // TODO
-			// TODO we should have the nonce, except for wallet like metamask where it is not usre you get the nonce you start with
-		};
-		await env.savePendingExecution(pendingExecution);
-		return txHash;
-	}
-
-	async function executeByName<
-		TAbi extends Abi,
-		TFunctionName extends ContractFunctionName<TAbi, 'nonpayable' | 'payable'>,
-		TArgs extends ContractFunctionArgs<TAbi, 'nonpayable' | 'payable', TFunctionName> = ContractFunctionArgs<
-			TAbi,
-			'nonpayable' | 'payable',
-			TFunctionName
-		>
-	>(name: string, args: ExecutionArgs<TAbi, TFunctionName, TArgs>) {
-		const deployment = env.getOrNull<TAbi>(name);
-		if (!deployment) {
-			throw new Error(`no deployment named ${name}`);
-		}
-
-		return execute(deployment, args);
-	}
-
-	async function read<
-		TAbi extends Abi,
-		TFunctionName extends ContractFunctionName<TAbi, 'pure' | 'view'>,
-		TArgs extends ContractFunctionArgs<TAbi, 'pure' | 'view', TFunctionName> = ContractFunctionArgs<
-			TAbi,
-			'pure' | 'view',
-			TFunctionName
-		>
-	>(
-		deployment: Deployment<TAbi>,
-		args: ReadingArgs<TAbi, TFunctionName, TArgs>
-	): Promise<DecodeFunctionResultReturnType<TAbi, TFunctionName>> {
-		const {account, ...viemArgs} = args;
-		let address: `0x${string}` | undefined;
-		if (account) {
-			if (account.startsWith('0x')) {
-				address = account as `0x${string}`;
-			} else {
-				if (env.namedAccounts) {
-					address = env.namedAccounts[account];
-					if (!address) {
-						throw new Error(`no address for ${account}`);
-					}
-				} else {
-					throw new Error(`no accounts setup, cannot get address for ${account}`);
-				}
-			}
-		}
-
-		const artifactToUse = deployment as unknown as Artifact<TAbi>;
-		const abi = artifactToUse.abi;
-		const calldata = encodeFunctionData<TAbi, TFunctionName>({
-			abi,
-			functionName: viemArgs.functionName,
-			args: viemArgs.args,
-		} as any);
-
-		const result: `0x${string}` = (await env.network.provider.request({
-			method: 'eth_call',
-			params: [
-				{
-					to: deployment.address,
-					type: '0x2',
-					from: address,
-					chainId: `0x${env.network.chain.id.toString(16)}` as `0x${string}`,
-					data: calldata,
-					// value: `0x${viemArgs.value?.toString(16)}` as `0x${string}`,
-				},
-			] as any, // TODO fix eip-1193 ?,,
-		})) as `0x${string}`;
-
-		const parsed = decodeFunctionResult<TAbi, TFunctionName>({
-			abi,
-			functionName: viemArgs.functionName,
-			data: result,
-			args: viemArgs.args,
-		} as any);
-
-		return parsed as DecodeFunctionResultReturnType<TAbi, TFunctionName>;
-	}
-
-	async function readByName<
-		TAbi extends Abi,
-		TFunctionName extends ContractFunctionName<TAbi, 'pure' | 'view'>,
-		TArgs extends ContractFunctionArgs<TAbi, 'pure' | 'view', TFunctionName> = ContractFunctionArgs<
-			TAbi,
-			'pure' | 'view',
-			TFunctionName
-		>
-	>(
-		name: string,
-		args: ReadingArgs<TAbi, TFunctionName, TArgs>
-	): Promise<DecodeFunctionResultReturnType<TAbi, TFunctionName>> {
-		const deployment = env.getOrNull<TAbi>(name);
-		if (!deployment) {
-			throw new Error(`no deployment named ${name}`);
-		}
-
-		return read(deployment, args);
-	}
 	async function deploy<TAbi extends Abi>(
 		name: string,
 		args: DeploymentConstruction<TAbi>,
@@ -598,9 +328,5 @@ extendEnvironment((env: Environment) => {
 	}
 
 	env.deploy = deploy;
-	env.execute = execute;
-	env.executeByName = executeByName;
-	env.read = read;
-	env.readByName = readByName;
 	return env;
 });