+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/deploy
 
+## 1.0.0-next.4
+
+### Patch Changes
+
+- add @rocketh/read-execute
+- Updated dependencies
+  - rocketh@1.0.0-next.4
+
 ## 1.0.0-next.3
 
 ### Patch Changes