+++ packages/rocketh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rocketh
 
+## 1.0.0-next.4
+
+### Patch Changes
+
+- add @rocketh/read-execute
+
 ## 1.0.0-next.3
 
 ### Patch Changes